Expanded defect behaviors
All variables are currently global just to simplify the function args. Functions are abstracted away from main into cluster_dynamics.cpp. The primary issue right now is the size of certain calculations, leading to the size_t values maxing out. This could be due to an error in value scaling or we simply need to support larger values.